Removed uncessary code from the uplaod file
